linkify: Remove repeated code.
While we are at it, a bug became obvious: I was getting the list of files
that from the current directory *before* moving into the directory where the
dotfiles were, which potentially got an incorrect list of directories.

It is now hopefully fixed.

Thanks go to both Rik Poggi and Winston Ewert from
<http://codereview.stackexchange.com> for the assistance in improving this
code.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>